+++ locales/index.js
@@ -5,11 +5,15 @@ import ru from './ru';
 import es from './es';
 import pt_br from './pt_br';
 import sk from './sk';
+import it from './it';
+import ua from './ua';
 
 export default {
     en,
     ru,
     es,
     'pt-BR': pt_br,
-    sk
+    sk,
+    it,
+    ua
 };