+++ components/ProfileView/ProfileViewMobile.js
@@ -27,7 +27,7 @@ export default class ProfileViewMobile extends PureComponent<Props> {
                             Repos
                         </UIText>
                         <UIText>
-                            {user.public_repos}
+                            {user.repositories.totalCount}
                         </UIText>
                     </View>
                     <View style={styles.statsBlock}>
@@ -35,7 +35,7 @@ export default class ProfileViewMobile extends PureComponent<Props> {
                             Gists
                         </UIText>
                         <UIText>
-                            {user.public_gists}
+                            {user.gists.totalCount}
                         </UIText>
                     </View>
                     <View style={styles.statsBlock}>
@@ -43,7 +43,7 @@ export default class ProfileViewMobile extends PureComponent<Props> {
                             Followers
                         </UIText>
                         <UIText>
-                            {user.followers}
+                            {user.followers.totalCount}
                         </UIText>
                     </View>
                     <View style={styles.statsBlock}>
@@ -51,7 +51,7 @@ export default class ProfileViewMobile extends PureComponent<Props> {
                             Following
                         </UIText>
                         <UIText>
-                            {user.following}
+                            {user.following.totalCount}
                         </UIText>
                     </View>
                 </View>
@@ -87,9 +87,9 @@ export default class ProfileViewMobile extends PureComponent<Props> {
                         ) : null
                     }
                     {
-                        user.blog ? (
+                        user.websiteUrl ? (
                             <TextIcon
-                                text={user.blog}
+                                text={user.websiteUrl}
                                 iconName={'link'}
                             />
                         ) : null